re-initializing backend scaffold
