+++ Backend_Application/src/main/java/dsd/cohort/application/Application.java
@@ -1,0 +1,13 @@
+package dsd.cohort.application;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Application.class, args);
+	}
+
+}
