+++ Backend_Application/src/main/java/dsd/cohort/application/config/CORSConfig.java
@@ -1,4 +1,4 @@
-package dsdcohort.Application.config;
+package dsd.cohort.application.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
