+++ Backend_Application/src/main/java/dsd/cohort/application/testController.java
@@ -1,0 +1,15 @@
+package dsd.cohort.application;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/")
+public class testController {
+
+    @GetMapping("/")
+    public String test() {
+        return "this is a test";
+    }
+}
