--- Backend_Application/src/main/java/dsdcohort/Application/ServletInitializer.java
@@ -1,13 +1,0 @@
-package dsdcohort.Application;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(Application.class);
-	}
-
-}
