+++ Backend_Application/src/main/resources/application.properties
@@ -1,8 +1,11 @@
 server.port=8080
 
-spring.thymeleaf.cache=false
-spring.thymeleaf.enabled=true
-spring.thymeleaf.prefix=classpath:/templates/
-spring.thymeleaf.suffix=.html
+spring.application.name=application
 
-spring.application.name=Spring Boot
+spring.h2.console.enabled=true
+spring.datasource.url=jdbc:h2:mem:test
+spring.jpa.hibernate.ddl-auto=none
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
