+++ Backend_Application/src/test/java/dsd/cohort/application/ApplicationTests.java
@@ -1,0 +1,13 @@
+package dsd.cohort.application;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
