--- Backend_Application/src/test/java/dsdcohort/Application/ApplicationTests.java
@@ -1,13 +1,0 @@
-package dsdcohort.Application;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
